MDEXP-628 - move s3client to configuration
