+++ src/test/java/org/folio/dataexp/service/SlicerProcessorTest.java
@@ -8,7 +8,7 @@ import org.folio.dataexp.domain.entity.JobExecutionExportFilesStatus;
 import org.folio.dataexp.repository.ExportIdEntityRepository;
 import org.folio.dataexp.repository.JobExecutionEntityRepository;
 import org.folio.dataexp.repository.JobExecutionExportFilesEntityRepository;
-import org.folio.dataexp.service.export.storage.FolioS3ClientFactory;
+import org.folio.s3.client.FolioS3Client;
 import org.folio.spring.scope.FolioExecutionContextSetter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +25,7 @@ class SlicerProcessorTest extends BaseDataExportInitializer {
   private static final String UPLOADED_FILE_PATH_CQL = "src/test/resources/upload_for_slicer.cql";
 
   @Autowired
-  private FolioS3ClientFactory folioS3ClientFactory;
+  private FolioS3Client s3Client;
   @Autowired
   private InputFileProcessor inputFileProcessor;
   @Autowired
@@ -47,7 +47,6 @@ class SlicerProcessorTest extends BaseDataExportInitializer {
     fileDefinition.setUploadFormat(FileDefinition.UploadFormatEnum.CQL);
     fileDefinition.setJobExecutionId(UUID.randomUUID());
 
-    var s3Client = folioS3ClientFactory.getFolioS3Client();
     s3Client.createBucketIfNotExists();
 
     var path = String.format(PATTERN_TO_SAVE_FILE, fileDefinition.getId(), fileDefinition.getFileName());
